Make all Girder exceptions a subclass of GirderBaseException.
This moves `utility.path.NotFoundException` to
`exceptions.ResourcePathNotFound`, keeping it exposed as
`utility.path.NotFoundException`.

This moves `api.rest.RestException` to `exceptions.RestException`.  I
did not change all of the references to importing `RestException` to
import from the `exceptions` modules, as it is still accessible from the
`api.rest` module and this would be a fairly extensive change.  If
others think the import references should be changed, let me know.

I converted the `__init__` calls in `exceptions` to all use `super`
based on meatspace conversations.
